requirements.txt
aiofiles==25.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.25.2
hyperframe==6.1.0
idna==3.11
lyricsgenius==3.7.6
magic-filter==1.0.12
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.0
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==20.7
requests
telegraph
python-dotenv